[Bronze III] Title: 첫 글자를 대문자로, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ C++20/백준/Bronze/4458. 첫 글자를 대문자로/첫 글자를 대문자로.cc
@@ -13,7 +13,7 @@ int main() {
         getline(cin, s);
         for (int i = 0; i < s.length(); i++) {
             if (i == 0 && 97 <= s[i] && (int)s[i] <= 122) {
-                cout << (char)((int)s[i] - 32);
+                cout << (char)(s[i] - 32);
             }
             else
                 cout << s[i];
